+++ examples/example-simplexcontour3d.cxx
@@ -53,7 +53,7 @@ int main(void)
 
   double t=0;
   double dt=0.1;
-  int ii=0;
+  size_t ii=0;
   auto t0=std::chrono::system_clock::now();
   double i0=ii;
 
@@ -73,7 +73,7 @@ int main(void)
 
   while (ii<nspin)
   {
-    for (int ipoint=0, ival=0;ipoint<points.size(); ipoint+=3,ival++)
+    for (size_t ipoint=0, ival=0;ipoint<points.size(); ipoint+=3,ival++)
       values[ival]=G(points[ipoint+0],points[ipoint+1],points[ipoint+2],t);
