+++ examples/example-surf2d.cxx
@@ -44,7 +44,7 @@ int main(const int argc, const char *argv[])
 
   double t=0;
   double dt=0.1;
-  int ii=0;
+  size_t ii=0;
   auto t0=std::chrono::system_clock::now();
   double i0=ii;
